gh-214: Splits up and Nodifies many synth-related tests.
