+++ tests/unit/all-tests.html
@@ -44,6 +44,11 @@
         "./html/scheduler-tests.html",
         "./html/scheduling-ugen-tests.html",
         "./html/synth-tests.html",
+        "./html/synth-environment-tests.html",
+        "./html/synth-group-tests.html",
+        "./html/synth-instantiation-tests.html",
+        "./html/synth-note-target-tests.html",
+        "./html/synth-scheduled-tests.html",
         "./html/trigger-ugen-tests.html",
         "./html/ugen-tests.html",
         "./html/webaudio-tests.html",
