+++ tests/unit/html/synth-group-tests.html
@@ -2,7 +2,7 @@
 <html>
     <head>
         <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-        <title>Synth Unit Tests</title>
+        <title>Synth Group Unit Tests</title>
 
         <link rel="stylesheet" media="screen" href="../../../node_modules/infusion/tests/lib/qunit/css/qunit.css" />
 
@@ -41,11 +41,11 @@
         <script src="../../../node_modules/infusion/tests/test-core/utils/js/IoCTestUtils.js"></script>
 
         <script src="../js/flocking-test-utils.js"></script>
-        <script src="../js/synth-tests.js"></script>
+        <script src="../js/synth-group-tests.js"></script>
     </head>
 
     <body id="body">
-        <h1 id="qunit-header">Synth Unit Tests</h1>
+        <h1 id="qunit-header">Synth Group Unit Tests</h1>
         <h2 id="qunit-banner"></h2>
         <div id="qunit-testrunner-toolbar"></div>
         <h2 id="qunit-userAgent"></h2>
