+++ tests/unit/html/synth-scheduled-tests.html
@@ -2,7 +2,7 @@
 <html>
     <head>
         <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-        <title>Synth Unit Tests</title>
+        <title>Scheduled Synth Unit Tests</title>
 
         <link rel="stylesheet" media="screen" href="../../../node_modules/infusion/tests/lib/qunit/css/qunit.css" />
 
@@ -31,21 +31,17 @@
 
         <script src="../../../src/ugens/core.js"></script>
         <script src="../../../src/ugens/oscillators.js"></script>
-        <script src="../../../src/ugens/envelopes.js"></script>
-        <script src="../../../src/ugens/random.js"></script>
-        <script src="../../../src/ugens/math.js"></script>
-        <script src="../../../src/ugens/scheduling.js"></script>
 
         <script src="../../../node_modules/infusion/tests/lib/qunit/js/qunit.js"></script>
         <script src="../../../node_modules/infusion/tests/test-core/jqUnit/js/jqUnit.js"></script>
         <script src="../../../node_modules/infusion/tests/test-core/utils/js/IoCTestUtils.js"></script>
 
         <script src="../js/flocking-test-utils.js"></script>
-        <script src="../js/synth-tests.js"></script>
+        <script src="../js/synth-scheduled-tests.js"></script>
     </head>
 
     <body id="body">
-        <h1 id="qunit-header">Synth Unit Tests</h1>
+        <h1 id="qunit-header">Scheduled Synth Unit Tests</h1>
         <h2 id="qunit-banner"></h2>
         <div id="qunit-testrunner-toolbar"></div>
         <h2 id="qunit-userAgent"></h2>
